Add QgsStackTrace class
