+++ src/app/qgscrashhandler.cpp
@@ -21,104 +21,19 @@
 #include "qgsproject.h"
 #include "qgscrashdialog.h"
 #include "qgscrashreport.h"
+#include "qgsstacktrace.h"
 
 #ifdef Q_OS_WIN
 LONG WINAPI QgsCrashHandler::handle( struct _EXCEPTION_POINTERS *ExceptionInfo )
 {
-  HANDLE process = GetCurrentProcess();
-  // TOOD Pull symbols from symbol server.
-  // TOOD Move this logic to generic stack trace class to handle each
-  // platform.
-  SymInitialize( process, NULL, TRUE );
-
-  // StackWalk64() may modify context record passed to it, so we will
-  // use a copy.
-  CONTEXT context_record = *ExceptionInfo->ContextRecord;
-  // Initialize stack walking.
-  STACKFRAME64 stack_frame;
-  memset( &stack_frame, 0, sizeof( stack_frame ) );
-#if defined(_WIN64)
-  int machine_type = IMAGE_FILE_MACHINE_AMD64;
-  stack_frame.AddrPC.Offset = context_record.Rip;
-  stack_frame.AddrFrame.Offset = context_record.Rbp;
-  stack_frame.AddrStack.Offset = context_record.Rsp;
-#else
-  int machine_type = IMAGE_FILE_MACHINE_I386;
-  stack_frame.AddrPC.Offset = context_record.Eip;
-  stack_frame.AddrFrame.Offset = context_record.Ebp;
-  stack_frame.AddrStack.Offset = context_record.Esp;
-#endif
-  stack_frame.AddrPC.Mode = AddrModeFlat;
-  stack_frame.AddrFrame.Mode = AddrModeFlat;
-  stack_frame.AddrStack.Mode = AddrModeFlat;
-
-  SYMBOL_INFO *symbol = ( SYMBOL_INFO * ) qgsMalloc( sizeof( SYMBOL_INFO ) + MAX_SYM_NAME );
-  symbol->SizeOfStruct = sizeof( SYMBOL_INFO );
-  symbol->MaxNameLen = MAX_SYM_NAME;
-
-  IMAGEHLP_LINE *line = ( IMAGEHLP_LINE * ) qgsMalloc( sizeof( IMAGEHLP_LINE ) );
-  line->SizeOfStruct = sizeof( IMAGEHLP_LINE );
-
-  IMAGEHLP_MODULE *module = ( IMAGEHLP_MODULE * ) qgsMalloc( sizeof( IMAGEHLP_MODULE ) );
-  module->SizeOfStruct = sizeof( IMAGEHLP_MODULE );
-
-  QList<QgsCrashReport::StackLine> stack;
-  while ( StackWalk64( machine_type,
-                       GetCurrentProcess(),
-                       GetCurrentThread(),
-                       &stack_frame,
-                       &context_record,
-                       NULL,
-                       &SymFunctionTableAccess64,
-                       &SymGetModuleBase64,
-                       NULL ) )
-  {
-
-    DWORD64 displacement = 0;
-
-    if ( SymFromAddr( process, ( DWORD64 )stack_frame.AddrPC.Offset, &displacement, symbol ) )
-    {
-      DWORD dwDisplacement;
-      QString fileName;
-      QString lineNumber;
-      QString moduleName;
-      if ( SymGetLineFromAddr( process, ( DWORD )( stack_frame.AddrPC.Offset ), &dwDisplacement, line ) )
-      {
-        fileName = QString( line->FileName );
-        lineNumber = QString::number( line->LineNumber );
-      }
-      else
-      {
-        fileName = "(unknown file)";
-        lineNumber = "(unknown line)";
-      }
-      if ( SymGetModuleInfo( process, ( DWORD )( stack_frame.AddrPC.Offset ), module ) )
-      {
-        moduleName = QString( module->ModuleName );
-      }
-      else
-      {
-        moduleName = "(unknown module)";
-      }
-      QgsCrashReport::StackLine stackline;
-      stackline.moduleName = moduleName;
-      stackline.fileName = fileName;
-      stackline.lineNumber = lineNumber;
-      stackline.symbolName = QString( symbol->Name );
-      stack.append( stackline );
-    }
-  }
-
-  qgsFree( symbol );
-  qgsFree( line );
-  qgsFree( module );
-
+  QVector<QgsStackTrace::StackLine> stack = QgsStackTrace::trace( ExceptionInfo );
   showCrashDialog( stack );
 
   return EXCEPTION_EXECUTE_HANDLER;
 }
+#endif
 
-void QgsCrashHandler::showCrashDialog( const QList<QgsCrashReport::StackLine> &stack )
+void QgsCrashHandler::showCrashDialog( const QVector<QgsStackTrace::StackLine> &stack )
 {
 
   QgsCrashDialog dlg( QApplication::activeWindow() );
@@ -141,4 +56,3 @@ void QgsCrashHandler::restartApplication()
   QProcess::startDetached( path, arguments, QDir::toNativeSeparators( QCoreApplication::applicationDirPath() ) );
 
 }
-#endif
