+++ src/core/qgsstacktrace.cpp
@@ -1,0 +1,153 @@
+/***************************************************************************
+  qgsstacktrace.cpp - QgsStackTrace
+
+ ---------------------
+ begin                : 24.4.2017
+ copyright            : (C) 2017 by Nathan Woodrow
+ email                : woodrow.nathan@gmail.com
+ ***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+#include "qgsstacktrace.h"
+
+#include <QVector>
+
+#ifdef WIN32
+#include <windows.h>
+#include <dbghelp.h>
+#endif
+
+#include "qgis.h"
+
+#ifdef Q_OS_WIN
+QVector<QgsStackTrace::StackLine> QgsStackTrace::trace(_EXCEPTION_POINTERS *ExceptionInfo)
+{
+  QVector<QgsStackTrace::StackLine> stack;
+#ifndef QGISDEBUG
+  return stack;
+#endif
+
+  HANDLE process = GetCurrentProcess();
+  // TOOD Pull symbols from symbol server.
+  SymInitialize( process, NULL, TRUE );
+
+  // StackWalk64() may modify context record passed to it, so we will
+  // use a copy.
+  CONTEXT context_record = *ExceptionInfo->ContextRecord;
+  // Initialize stack walking.
+  STACKFRAME64 stack_frame;
+  memset( &stack_frame, 0, sizeof( stack_frame ) );
+#if defined(_WIN64)
+  int machine_type = IMAGE_FILE_MACHINE_AMD64;
+  stack_frame.AddrPC.Offset = context_record.Rip;
+  stack_frame.AddrFrame.Offset = context_record.Rbp;
+  stack_frame.AddrStack.Offset = context_record.Rsp;
+#else
+  int machine_type = IMAGE_FILE_MACHINE_I386;
+  stack_frame.AddrPC.Offset = context_record.Eip;
+  stack_frame.AddrFrame.Offset = context_record.Ebp;
+  stack_frame.AddrStack.Offset = context_record.Esp;
+#endif
+  stack_frame.AddrPC.Mode = AddrModeFlat;
+  stack_frame.AddrFrame.Mode = AddrModeFlat;
+  stack_frame.AddrStack.Mode = AddrModeFlat;
+
+  SYMBOL_INFO *symbol = ( SYMBOL_INFO * ) qgsMalloc( sizeof( SYMBOL_INFO ) + MAX_SYM_NAME );
+  symbol->SizeOfStruct = sizeof( SYMBOL_INFO );
+  symbol->MaxNameLen = MAX_SYM_NAME;
+
+  IMAGEHLP_LINE *line = ( IMAGEHLP_LINE * ) qgsMalloc( sizeof( IMAGEHLP_LINE ) );
+  line->SizeOfStruct = sizeof( IMAGEHLP_LINE );
+
+  IMAGEHLP_MODULE *module = ( IMAGEHLP_MODULE * ) qgsMalloc( sizeof( IMAGEHLP_MODULE ) );
+  module->SizeOfStruct = sizeof( IMAGEHLP_MODULE );
+
+  while ( StackWalk64( machine_type,
+                       GetCurrentProcess(),
+                       GetCurrentThread(),
+                       &stack_frame,
+                       &context_record,
+                       NULL,
+                       &SymFunctionTableAccess64,
+                       &SymGetModuleBase64,
+                       NULL ) )
+  {
+
+    DWORD64 displacement = 0;
+
+    if ( SymFromAddr( process, ( DWORD64 )stack_frame.AddrPC.Offset, &displacement, symbol ) )
+    {
+      DWORD dwDisplacement;
+      QString fileName;
+      QString lineNumber;
+      QString moduleName;
+      if ( SymGetLineFromAddr( process, ( DWORD )( stack_frame.AddrPC.Offset ), &dwDisplacement, line ) )
+      {
+        fileName = QString( line->FileName );
+        lineNumber = QString::number( line->LineNumber );
+      }
+      else
+      {
+        fileName = "(unknown file)";
+        lineNumber = "(unknown line)";
+      }
+      if ( SymGetModuleInfo( process, ( DWORD )( stack_frame.AddrPC.Offset ), module ) )
+      {
+        moduleName = QString( module->ModuleName );
+      }
+      else
+      {
+        moduleName = "(unknown module)";
+      }
+      QgsStackTrace::StackLine stackline;
+      stackline.moduleName = moduleName;
+      stackline.fileName = fileName;
+      stackline.lineNumber = lineNumber;
+      stackline.symbolName = QString( symbol->Name );
+      stack.append( stackline );
+    }
+  }
+
+  qgsFree( symbol );
+  qgsFree( line );
+  qgsFree( module );
+  return stack;
+
+}
+#endif
+
+QVector<QgsStackTrace::StackLine> QgsStackTrace::trace( unsigned int maxFrames )
+{
+  QVector<QgsStackTrace::StackLine> stack;
+#ifndef QGISDEBUG
+  return stack;
+#endif
+
+#ifdef Q_OS_LINUX
+  // TODO Add linux stack trace support. Pull it from main.cpp
+#endif
+  return stack;
+}
+
+QgsStackTrace::QgsStackTrace()
+{
+
+}
+
+bool QgsStackTrace::StackLine::isQgisModule() const
+{
+  return moduleName.toLower().contains( "qgis" );
+}
+
+bool QgsStackTrace::StackLine::isValid() const
+{
+  return !( fileName.toLower().contains( "exe_common" ) ||
+            fileName.toLower().contains( "unknown" ) ||
+            lineNumber.toLower().contains( "unknown" ) );
+
+}
